pyproject.toml
[project]
name = "roboot"
version = "0.1.0"
description = "Personal AI agent hub — voice, terminal, Claude Code management"
requires-python = ">=3.11"
dependencies = [
    "arcana-agent[all-providers]>=0.4.0",
    "pyyaml>=6.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
]

[project.optional-dependencies]
voice = [
    "SpeechRecognition>=3.10.0",
    "pyaudio>=0.2.14",
]
vision = [
    "opencv-python>=4.9.0",
]
telegram = [
    "python-telegram-bot>=21.0",
]
desktop = [
    "pywebview>=5.0",
]
all = [
    "roboot[voice,vision,telegram,desktop]",
]

[project.scripts]
roboot = "run:main"
